backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# LLM & RAG
langchain==0.1.0
langgraph==0.0.20
openai==1.10.0
sentence-transformers==2.3.1
transformers==4.36.2

# Vector & Graph
faiss-cpu==1.7.4
networkx==3.2.1
python-louvain==0.16

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
markdown==3.5.2
pdfplumber==0.10.3

# Data & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
redis==5.0.1
aioredis==2.0.1

# Multi-Modal
pytesseract==0.3.10
paddleocr==2.7.0
openai-whisper==20231117
pillow==10.2.0
opencv-python==4.9.0.80

# NLP & Analysis
spacy==3.7.2
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0

# External KB Integration
wikipedia-api==0.6.0
biopython==1.83

# Monitoring & Visualization
prometheus-client==0.19.0
plotly==5.18.0
seaborn==0.13.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Evaluation
ragas==0.1.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
